filter out all email addresses that don't fit our search criteria
This greatly reduces the number of false positives. Not all fetched addresses fit our search criteria, e.g. when there are multiple email addresses in the to: or when the searched email address is in the 'cc' header (we only look at 'to' for now), so only examine those addresses where the search criteria is a case insensitive substring of the full name + mail address string

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>