fix notmuch_database_open call
In the latest version of notmuch in git, notmuch_database_open returns
a status and takes its old return value as a reference parameter.
This patch adjusts code to pass the db pointer in a parameter and
accept the status as return value.  We don't do anything with the
status at present.